policy/k8s: Move PolicyWatcher into separate file
The policy watcher is shared for all watched resources. Therefore move
it out of the CNP file into a separate file. This commit contians no
changes to the function body.
